hooks/fix-systemd-generators-cryptroot: new hook
When using "root on zfs on dmcrypt" systemd is unaware that
the crypt devices for the root pool are still needed on shutdown.

So systemd hangs on shutdown, as it cannot shutdown devices if ZFS is
still using it.

This hook prevents systemd from managing dmcrypt devices if
using "root on zfs on dmcrypt".